requirements.txt
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
requests>=2.31.0
openpyxl>=3.1.0
gspread>=6.1.0
google-auth>=2.29.0
